+++ src/Shared functions.h
@@ -1,7 +1,7 @@
 #include <arduino.h>
 
 String addTrailingSpaces(String text, int TotalLength = 16);
-double clip(double n, double lower, double upper);
+long clip(long n, long lower, long upper);
 double roundDepth(double digit, int precision = 0);
 double stringToDouble(char* str);
 int stringToInt(char* str);
